fix(components): rename components screens. #47
