+++ frontend/src/router.ts
@@ -2,16 +2,16 @@ import { createWebHistory, createRouter } from 'vue-router'
 
 import HomeView from './views/HomeView.vue'
 import PetsView from './views/PetsView.vue'
-import ComponentesLuis from './views/ComponentesLuis.vue'
-import ComponentesSergio from './views/ComponentesSergio.vue'
+import Componentes1 from './views/Componentes1.vue'
+import Componentes2 from './views/Componentes2.vue'
 import TemplateScreen1 from './views/TemplateScreen1.vue'
 import TemplateScreen2 from './views/TemplateScreen2.vue'
 
 const routes = [
   { path: '/', component: HomeView },
   { path: '/pets', component: PetsView},
-  { path: '/componentesLuis', component: ComponentesLuis},
-  { path: '/componentesSergio', component: ComponentesSergio},
+  { path: '/componentes1', component: Componentes1},
+  { path: '/componentes2', component: Componentes2},
   { path: '/templateScreen1', component: TemplateScreen1},
   { path: '/templateScreen2', component: TemplateScreen2}
 ]
